// Railway-Frontend/src/app/shared/models/bookButtonSchema.ts
export class bookButton {
    userName: string;
    trainNumber: string;
    trainName: string;
    fromStation: string;
    fromStationArrivalTime: string;
    fromStationDepartureTime: string;
    toStation:string;
    toStationArrivalTime:string;
    toStationDepartureTime:string;
    className: string;
    availableSeats: number;
    fare:number;
    date: string;
}